requirements.txt
djangorestframework-simplejwt
django-cors-headers
djangorestframework
drf-yasg
django-allauth
python-dotenv
django-dotenv
psycopg2